+++ tools/tiffdither.c
@@ -141,14 +141,14 @@ static int fsdither(TIFF *in, TIFF *out)
                 bit = 0x80;
             }
             if (!lastpixel)
-                thisptr[0] += v * 7 / 16;
+                thisptr[0] += (short)(v * 7 / 16);
             if (!lastline)
             {
                 if (j != 0)
-                    nextptr[-1] += v * 3 / 16;
-                *nextptr++ += v * 5 / 16;
+                    nextptr[-1] += (short)(v * 3 / 16);
+                *nextptr++ += (short)(v * 5 / 16);
                 if (!lastpixel)
-                    nextptr[0] += v / 16;
+                    nextptr[0] += (short)(v / 16);
             }
         }
         if (TIFFWriteScanline(out, outline, i, 0) < 0)
@@ -206,14 +206,14 @@ static int processCompressOptions(char *opt)
     {
         char *cp = strchr(opt, ':');
         if (cp)
-            predictor = atoi(cp + 1);
+            predictor = (uint16_t)atoi(cp + 1);
         compression = COMPRESSION_LZW;
     }
     else if (strneq(opt, "zip", 3))
     {
         char *cp = strchr(opt, ':');
         if (cp)
-            predictor = atoi(cp + 1);
+            predictor = (uint16_t)atoi(cp + 1);
         compression = COMPRESSION_ADOBE_DEFLATE;
     }
     else
